Add ability to store previous request in a json database
card-counter will store files in `~/.card-counter`. This means a user
doesn't always have to enter in information such as API_TOKEN or
API_KEY and maybe a standard filter by option.

This will also store a `history.json` file as a lightweight database,
to help track and display the progress of boards over time.

Implement a local database as a json file

Change name from history to database