Create a means to generate a history file
card-counter will store files in `~/.card-counter`. This means a user
doesn't always have to enter in information such as API_TOKEN or
API_KEY and maybe a standard filter by option.

This will also store a `history.json` file as a lightweight database,
to help track and display the progress of boards over time.