Improve Error messaging
This looks at improving the way we handle and generate error messages
by using error error_chain!. This make it easy to solidify and handle
errors under one type. The eventual goal is to improve error messages
throughout the system to make the messages actionable to users.

This is referenced and kept track in #2.