Refactor decks and deck filtering
This refactor now saves scoring of all lists in a Trello boards. This
was added to allow reanalysis of old data. Say if a user originally
filters a set of lists for a board on day 1, but then wants a detailed
scoring for all lists on day 5, we will now be able to compare all
decks. Previously we would have only been able to return information
for the subset of decks that matched the filters from day 1.