feat: Add support saving to AWS
This adds the ability for card-counter to save and pull entries from
DyanamoDB

As well there was a Database struct and traits for database added to
more easily abstract over saving and retrieving items from different
types of databases. This is first expanded on with DynamoDB but plans
are to use it for all other current and future database systems.