Experiment with using plotters as a graphics library
By using plotter we can expand our chart output to be able to output
to SVG and in the terminal itself.

By enhancing our charting ability that demands that we add a new
option to the burndown, "output", which allows the user to choose how
the burndown calculation is displayed. The default option is CSV, the
one we support currently. The other two options are
- terminal: so we print a line plot in the terminal
- SVG: print an SVG so the image may be portable and be imported into
other systems

Begin adding ascii backend

Add experimental backend

Save all current work with plotters

ToDrop: Update cargo.lock file